requirements.txt
# PRODUCTION REQUIREMENTS - Legal Intelligence Assistant
# Optimized for Local PC, Kaggle, Hugging Face Spaces
# Total size: ~3.5 GB | RAM needed: 8GB+

# ========================================
# CORE ML/NLP (Pinned versions)
# ========================================
torch==2.1.2
transformers==4.36.2
sentence-transformers==2.2.2
torch-geometric==2.4.0

# ========================================
# DOCUMENT PROCESSING (Your excellent parser)
# ========================================
pdfplumber==0.10.3
PyPDF2==4.0.1
pytesseract==0.3.10
Pillow==10.1.0
python-docx==1.1.0

# ========================================
# NLP & TEXT PROCESSING
# ========================================
spacy==3.7.2
nltk==3.8.1

# ========================================
# VECTOR SEARCH & RETRIEVAL
# ========================================
faiss-cpu==1.8.0
rank-bm25==0.2.2

# ========================================
# WEB FRAMEWORK
# ========================================
streamlit==1.28.1

# ========================================
# DATA PROCESSING
# ========================================
pandas==2.1.4
numpy==1.24.3
scikit-learn==1.3.2

# ========================================
# VISUALIZATION
# ========================================
plotly==5.18.0
networkx==3.2

# ========================================
# UTILITIES
# ========================================
tqdm==4.66.1
python-dotenv==1.0.0
pyyaml==6.0.1



# # Core ML/NLP Libraries
# torch>=2.0.0
# torchvision>=0.15.0
# torchaudio>=2.0.0
# transformers>=4.30.0
# sentence-transformers>=2.2.0
# torch-geometric>=2.3.0

# # Document Processing
# pdfplumber>=0.9.0
# PyPDF2>=3.0.0
# pytesseract>=0.3.10
# python-docx>=0.8.11
# Pillow>=10.0.0

# # NLP & Text Processing
# spacy>=3.5.0
# nltk>=3.8.0

# # Vector Search & Retrieval
# faiss-cpu>=1.7.4
# rank-bm25>=0.2.2

# # LLM Framework
# langchain>=0.0.200
# langchain-community>=0.0.10

# # Web Framework
# streamlit>=1.25.0

# # Data Processing
# pandas>=2.0.0
# numpy>=1.24.0
# scikit-learn>=1.3.0

# # Visualization
# matplotlib>=3.7.0
# seaborn>=0.12.0
# plotly>=5.14.0
# networkx>=3.1.0

# # Evaluation
# datasets>=2.13.0
# evaluate>=0.4.0

# # Utilities
# tqdm>=4.65.0
# python-dotenv>=1.0.0
# pyyaml>=6.0
